Added getDifficulty function and helpers
Small test suite added for `getPartySize` function that returns `Maybe PartySize`.

I fought with this for a while, mostly trying to express the logic in a functional style without my usual imperative fallbacks. This code could probably be cleaned up quite a bit, once I get used to the APIs and idioms.